# public/api-docs.yaml
openapi: 3.0.0
info:
  title: 'PharmaControl API'
  description: "ERP para farmacias multi-sucursal en México con regulación COFEPRIS.\nAutenticación JWT propio (HS256). Todos los endpoints protegidos requieren Authorization: Bearer <token>."
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080/api/v1/'
    description: 'Local Docker'
paths:
...
  /v1/catalog/locations:
    get:
      tags:
        - 'Catalog · Locations'
      summary: 'Lista plana de ubicaciones'
      operationId: aa48b34a91caa0139e9d289a8aa274f2
      parameters:
        -
          name: level
          in: query
          required: false
          schema:
            type: integer
            enum:
              - 1
              - 2
              - 3
              - 4
        -
          name: parent_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        -
          name: is_active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: 'Lista de ubicaciones'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/LocationResponse' } }
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - 'Catalog · Locations'
      summary: 'Crear ubicación'
      description: 'Crea un nodo en la jerarquía. El nivel debe coincidir con el nivel hijo del padre dado. Las Zonas (level=1) deben tener parent_id nulo.'
      operationId: b0877652754f44d72ac3e303e4a0ec11
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - level
              properties:
                name:
                  type: string
                  example: 'Pasillo A'
                  maxLength: 100
                level:
                  type: integer
                  example: 2
                  enum: [1, 2, 3, 4]
                parent_id:
                  type: string
                  format: uuid
                  nullable: true
                description:
                  type: string
                  nullable: true
                  maxLength: 255
              type: object
      responses:
        '201':
          description: 'Ubicación creada'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 'parent_id no encontrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 'Nombre duplicado en el mismo nivel'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'Nivel inválido para el padre dado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
  /v1/catalog/locations/tree:
    get:
      tags:
        - 'Catalog · Locations'
      summary: 'Árbol completo de ubicaciones'
      operationId: e1e9c9bc0306a20f03e0000f33bfea92
      parameters:
        -
          name: is_active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: 'Árbol anidado Zona > Pasillo > Estante > Posición'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/LocationResponse' } }
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        -
          bearerAuth: []
  /v1/catalog/locations/leaves:
    get:
      tags:
        - 'Catalog · Locations'
      summary: 'Posiciones activas (level=4)'
      description: 'Devuelve solo las Posiciones activas. Se usa en el formulario de creación de productos.'
      operationId: efad3dd3c637b6ee8d7502869f7c7c4d
      parameters:
        -
          name: is_active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: 'Lista de posiciones'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/LocationResponse' } }
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        -
          bearerAuth: []
  '/v1/catalog/locations/{id}':
    get:
      tags:
        - 'Catalog · Locations'
      summary: 'Obtener ubicación por ID'
      operationId: fa0625a64dc779db918da8acfd166f93
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'Ubicación encontrada'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - 'Catalog · Locations'
      summary: 'Actualizar ubicación'
      description: 'Actualiza nombre y descripción. El parent_id y el level son inmutables tras la creación.'
      operationId: bc9bb9fce46af89e12f10d361a730e0d
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  nullable: true
                  maxLength: 255
              type: object
      responses:
        '200':
          description: 'Ubicación actualizada'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'Nombre duplicado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - 'Catalog · Locations'
      summary: 'Desactivar ubicación'
      description: 'Soft delete lógico (is_active = false). Los hijos NO se desactivan automáticamente.'
      operationId: ad515c6e4b47d1e14e9d6296f22dadb5
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    LocationResponse:
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Pos 1'
        level:
          description: '1=Zona, 2=Pasillo, 3=Estante, 4=Posición'
          type: integer
          example: 4
        level_label:
          type: string
          example: Posición
        parent_id:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
          example: true
        is_leaf:
          type: boolean
          example: true
        full_path:
          type: string
          example: 'OTC General > Pasillo 1 > Estante 1 > Pos 1'
        created_by:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...
    ErrorResponse:
      properties:
        message:
          type: string
          example: 'Recurso no encontrado.'
        error:
          type: string
          example: NOT_FOUND
          nullable: true
      type: object
...
  responses:
    Unauthorized:
      description: 'Token ausente, inválido o expirado'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Unauthenticated.
    Forbidden:
      description: 'Sin el permiso requerido'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 'Recurso no encontrado'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NoContent:
      description: 'Operación exitosa sin contenido'
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...